Add support for wildcards through module.exports.
This is basically an implementation of and closes #1. It also bumps the
version to 0.2.0, updates the documentation and adds tests for the
feature.

The option may be disabled by setting `options.moduleWildcards` to
`false`. Since it shouldn't cause any harm to people not using
`EventEmitter2` (except in really strange cases), it doesn't seem like
an issue to enable it by default.